requirements.txt
streamlit
pandas
requests
python-dotenv
yfinance
matplotlib
fpdf
forex-python
prophet
plotly